+++ src/lib/entity/detail/sz-entity-detail-graph/sz-entity-detail-graph.component.html
@@ -32,7 +32,7 @@
     [buildOut]="buildOut"
     [linkGravity]="5"
     [highlight]="entityNodeColors"
-    [filter]="entityNodeFilterByDataSource"
+    [filter]="entityNodeFilters"
     [showLinkLabels]="_showMatchKeys"
     [captureMouseWheel]="captureMouseWheel"
     (onWheelScroll)="onGraphScrollEvent($event)"
